Add files via upload
A system that can easily handle all the necessary and minute information, as well as sufficient database security according to the user, is needed for gym management. Software is needed for them to store member data.
The graphical user interface (GUI) tracks attendance, reports, pending payments, customer reports, inquiry details, receipts for members, and other activities that take place in the gym and during lock-up.
This project is created to help a gym and fitness centre automate its record-keeping processes and save them in the form of a sizable and user-friendly database, making it even easier for the staff to access.